requirements.txt
python-dotenv
fastapi[all]
inewave